backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
pydantic==2.5.3
google-api-python-client==2.121.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-groq>=0.1.6
python-dotenv==1.0.0
cryptography==42.0.0
prometheus-client==0.19.0
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Development
python-multipart==0.0.6
requests==2.31.0
